refactor: UserServiceHandler, UserServiceClient - user-service로 요청하는 클래스 분리 및 패키지 domain의 상위로 이동
