+++ src/main/java/com/polar_moviechart/movieservice/controller/secureapi/MovieControllerSecure.java
@@ -2,6 +2,7 @@ package com.polar_moviechart.movieservice.controller.secureapi;
 
 import com.polar_moviechart.movieservice.domain.service.movie.MovieCommandService;
 import com.polar_moviechart.movieservice.domain.service.movie.MovieQueryService;
+import com.polar_moviechart.movieservice.handler.UserServiceHandler;
 import com.polar_moviechart.movieservice.utils.CustomResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -16,11 +17,13 @@ import org.springframework.web.bind.annotation.*;
 public class MovieControllerSecure {
     private final MovieCommandService movieCommandService;
     private final MovieQueryService movieQueryService;
+    private final UserServiceHandler userServiceHandler;
 
     @GetMapping("/{code}/rating")
     public ResponseEntity<CustomResponse<Double>> getMovieRating(HttpServletRequest request,
                                        @PathVariable(name = "code") int code) {
         Long userId = (Long) request.getAttribute("userId");
+        userServiceHandler.validateUserExists(userId);
         Double movieRating = movieQueryService.getUserMovieRating(code, userId);
 
         return ResponseEntity.ok(new CustomResponse<>(movieRating));
@@ -30,10 +33,14 @@ public class MovieControllerSecure {
     public ResponseEntity<CustomResponse<Double>> updateRating(HttpServletRequest request,
                                        @PathVariable(name = "code") int code,
                                        @RequestBody UpdateRatingRequest updateRatingRequest) {
-        String userIdString = request.getHeader("X-User-Id");
-        Long userId = Long.valueOf(userIdString);
+        Long userId = getUserId(request);
+        userServiceHandler.validateUserExists(userId);
         double ratingValue = movieCommandService.updateRating(code, userId, updateRatingRequest);
 
         return ResponseEntity.ok(new CustomResponse<>(ratingValue));
+    }
+
+    private Long getUserId(HttpServletRequest request) {
+        return Long.parseLong(request.getHeader("X-User-Id"));
     }
 }
