+++ src/main/java/com/polar_moviechart/movieservice/domain/service/movie/MovieCommandService.java
@@ -2,7 +2,6 @@ package com.polar_moviechart.movieservice.domain.service.movie;
 
 import com.polar_moviechart.movieservice.controller.secureapi.UpdateRatingRequest;
 import com.polar_moviechart.movieservice.domain.entity.MovieRating;
-import com.polar_moviechart.movieservice.domain.service.UserValidationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,7 +14,6 @@ public class MovieCommandService {
     private final MovieQueryService movieQueryService;
     private final MovieRatingCommandService movieRatingCommandService;
     private final MovieRatingQueryService movieRatingQueryService;
-    private final UserValidationService userValidationService;
 
     @Transactional
     public double updateRating(Integer code, Long userId, UpdateRatingRequest updateRatingRequest) {
@@ -26,7 +24,6 @@ public class MovieCommandService {
             MovieRating movieRating = movieRatingOptional.get();
             movieRating.setRating(ratingValue);
         } else {
-            userValidationService.validateUserExists(userId);
             movieQueryService.validateExists(code);
             movieRatingCommandService.updateRating(code, userId, ratingValue);
         }
